requirements.txt
# Python >= 3.10 required (developed with 3.11.9)

# ── Embedding（本地，完全免費） ──────────────────────────
sentence-transformers>=2.7.0

# ── Vector DB ────────────────────────────────────────────
chromadb>=0.5.3

# ── LLM 呼叫 ─────────────────────────────────────────────
# litellm>=1.40.0
openai>=1.0.0
# ── PDF 解析 ──────────────────────────────────────────────
pdfplumber>=0.11.0      # 優先：保留表格結構，適合法規文件
pypdf>=4.0.0            # 備援：輕量 PDF 解析

# ── 工具套件 ──────────────────────────────────────────────
python-dotenv>=1.0.0    # .env 環境變數載入
numpy>=1.24.0           # sentence-transformers 依賴

# ── 可選：加速 / 品質提升 ────────────────────────────────
torch>=2.0.0          # GPU 加速 embedding（選裝，CPU 版亦可用）
rank-bm25>=0.2.2      # BM25 混合檢索（進階功能）
